requirements.txt
flask==3.0.0
tweepy==4.14.0
textblob==0.17.1
python-dotenv==1.0.0